# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sistema de Soporte Técnico ISP — equipo ACC
  description: >
    Contrato REST de los 5 microservicios del PFC. Cada servicio corre en su
    propio puerto y mantiene su propia base de datos (auth_db / ticket_db /
    notifications_db / ai_db / report_db). Desde la Entrega 4 el punto de
    entrada único para los clientes (web y móvil) es el API Gateway en el
    puerto 8000, que reenvía cada prefijo /api/v1/<servicio>/** al
    microservicio correspondiente sin reescribir la ruta (ver
    services/api-gateway/src/main/resources/application.yml) -- los puertos
    directos de cada servicio siguen respondiendo (útiles para depurar un
    servicio en aislamiento) pero ya no son la vía real de acceso.
  version: "4.0.0"
  contact:
    name: Equipo ACC — Aplicaciones Distribuidas (ISR-701), UTEQ
servers:
  - url: http://localhost:8000/api/v1/auth
    description: auth-service, a través del API Gateway (Entrega 4, vía real)
  - url: http://localhost:8000/api/v1/tickets
    description: ticket-service, a través del API Gateway (Entrega 4, vía real)
  - url: http://localhost:8000/api/v1/notifications
    description: notification-service, a través del API Gateway (Entrega 4, vía real)
  - url: http://localhost:8000/api/v1/ai
    description: ai-service, a través del API Gateway (Entrega 4, vía real)
  - url: http://localhost:8000/api/v1/reports
    description: report-service, a través del API Gateway (Entrega 4, vía real)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      description: >
        Envoltura estándar de respuesta. auth-service, ticket-service y report-service (Java) la
        producen con la misma clase ApiResponse; notification-service (Node) construye a mano un
        objeto con la misma forma. ai-service (FastAPI) es la única excepción: devuelve el
        documento sin envolver (ver /classifications/{ticket_id}).
      properties:
        data: {}
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    Ticket:
      type: object
      properties:
        zone:
          type: string
          enum: [QUEVEDO_CENTRO, QUEVEDO_NORTE, QUEVEDO_SUR]
        ticketId:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
        technicianId:
          type: string
          format: uuid
          nullable: true
        category:
          type: string
          nullable: true
          enum: [CONECTIVIDAD, DNS, HARDWARE, CONFIGURACION, VELOCIDAD]
        priority:
          type: string
          nullable: true
          enum: [CRITICO, ALTO, MEDIO, BAJO]
        status:
          type: string
          enum: [NUEVO, ASIGNADO, EN_PROGRESO, ESCALADO, RESUELTO, CERRADO]
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        slaDeadline:
          type: string
          format: date-time
        slaBreached:
          type: boolean
        resolvedAt:
          type: string
          format: date-time
          nullable: true
...
    TicketSummaryResponse:
      type: object
      description: Fila del modelo de lectura CQRS (report-service), no el mismo objeto que Ticket.
      properties:
        zone:
          type: string
        ticketId:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
        technicianId:
          type: string
          format: uuid
          nullable: true
        category:
          type: string
        priority:
          type: string
        status:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
  responses:
...
    Forbidden:
      description: Rol autenticado sin permiso para esta operación
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponse"
paths:
...
  /{id}/status:
    patch:
      tags: [ticket-service]
      summary: Cambiar estado (TECNICO en su zona / ADMIN)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [NUEVO, ASIGNADO, EN_PROGRESO, ESCALADO, RESUELTO, CERRADO] }
                evidencePhotoBase64:
                  type: string
                  format: byte
                  nullable: true
                  description: >-
                    Foto de evidencia del cierre en sitio, codificada en Base64. Opcional: solo
                    la manda la app móvil cuando status=RESUELTO (Entregable 10 de la guía de
                    cierre); en cualquier otra transición el servicio la ignora aunque llegue.
                latitude:
                  type: number
                  format: double
                  nullable: true
                  description: Latitud GPS capturada al cerrar en sitio. Opcional, mismo caso que evidencePhotoBase64.
                longitude:
                  type: number
                  format: double
                  nullable: true
                  description: Longitud GPS capturada al cerrar en sitio. Opcional, mismo caso que evidencePhotoBase64.
      responses:
        "200":
          description: Estado actualizado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Ticket" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /{id}/assign:
    post:
      tags: [ticket-service]
      summary: Asignar técnico (TECNICO en su zona / ADMIN)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: technicianId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Técnico asignado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Ticket" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /tickets:
    get:
      tags: [report-service]
      summary: Lista filtrable del modelo de lectura (solo ADMIN)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: zone
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Lista de TicketSummaryResponse
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/TicketSummaryResponse" }
        "403": { $ref: "#/components/responses/Forbidden" }
...